fix delete lab and doc

+++ src/controllers/project.js
@@ -2,7 +2,8 @@
 
 const { Project, validateProject } = require("../models/project");
 const mongoose = require("mongoose");
-
+const SelectionController = require("./selection");
+const TeacherController = require("./teacher");
 /**
  * Get all projects
  * @returns {array} list of all projects
@@ -93,8 +94,20 @@ const update = async (id, { name, description, labId, selections }) =>
  * @param {string} id
  * @returns {object} project removed
  */
-const remove = async (id) =>
-  await Project.findByIdAndRemove(mongoose.Types.ObjectId(id));
+const remove = async (id) => {
+  const project = await Project.findById(mongoose.Types.ObjectId(id));
+
+  if (project) {
+    await Promise.all(
+      project.selections.map(
+        async (selectionId) => await SelectionController.remove(selectionId)
+      )
+    );
+    return await Project.findByIdAndRemove(mongoose.Types.ObjectId(id));
+  } else {
+    throw new Error("The project with the given ID was not found.");
+  }
+};
 
 /**
  * Validate project
